Fixed hanging bash commands from agent
If the agent puts out a command like
    'while true; do ls /root; sleep 1; done'
it will loop while also putting out an output (meaning the socket
doesn't timeout) so I've added a 30s cutoff. Without this it eventually
fails but counts as an incomplete task rather than just a fail so it
stops any of the overall stats working.